refactor: extract DisconnectNotice into abstract class OverlayNotice, implement another notice for Unauthorized errors
